+++ client/src/core.js
@@ -12,6 +12,7 @@ import 'jquery-contextmenu'
 import 'jquery-contextmenu/dist/jquery.ui.position'
 import 'jquery-contextmenu/dist/jquery.contextMenu.css'
 import '@/jquery.contextMenu.overrides.css'
+import './scss/bootstrap-theme.scss'
 // TODO: join dynamic form could be on any page - fix this
 import '@/join'