+++ src/Modules/Store/components/StoreList.test.js
@@ -1,0 +1,43 @@
+/* eslint-env mocha */
+
+import sinon from 'sinon'
+import { mount } from '@vue/test-utils'
+import { resetModules } from '>/utils'
+import StoreList from './StoreList'
+
+const assert = require('assert')
+
+function createMockStore () {
+  return {
+    added: '1983-04-10',
+    address: 'Tanja-Oswald-Ring 08c 281',
+    id: 15906,
+    name: 'betrieb_Bader Hammer KG',
+    region: 'Göttingen',
+    status: 3
+  }
+}
+
+describe('StoreList', () => {
+  const sandbox = sinon.createSandbox()
+
+  afterEach(() => {
+    sandbox.restore()
+    resetModules()
+  })
+
+  it('loads', () => {
+    assert(StoreList)
+  })
+
+  it('can render', () => {
+    const regionName = 'Test Region Name'
+    const wrapper = mount(StoreList, {
+      propsData: {
+        regionName,
+        stores: [createMockStore()]
+      }
+    })
+    assert.notEqual(wrapper.vm.$el.innerHTML.indexOf(regionName), -1)
+  })
+})