+++ tests/_support/Helper/Foodsharing.php
@@ -267,8 +267,13 @@ class Foodsharing extends \Codeception\Module\Db
 			'fetcher' => $this->faker->numberBetween(1, 8),
 		], $extra_params);
 
-		$id = $this->haveInDatabase('fs_abholzeiten', $params);
-		$params['id'] = $id;
+		try {
+			/* ToDo: Easy to generate a collision with the chosen randoms on big number of stores */
+			$id = $this->haveInDatabase('fs_abholzeiten', $params);
+			$params['id'] = $id;
+		} catch (\Exception $e) {
+			return $this->addRecurringPickup($store, $extra_params);
+		}
 
 		return $params;
 	}